+++ CHANGELOG.md
@@ -15,6 +15,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   you to either omit `functions` for nodes, which is common for the end node,
   or specify an empty function call list, if desired.
 
+### Deprecated
+
+- The `tts` parameter in `FlowManager.__init__()` is now deprecated and will
+  be removed in a future version. The `tts_say` action now pushes a
+  `TTSSpeakFrame`.
+
+### Fixed
+
+- Overhauled `pre_actions` and `post_actions` timing logic, making their timing more predictable and
+  eliminating some bugs. For example, now `tts_say` actions will always run after the bot response,
+  when used in `post_actions`.
+
 ## [0.0.17] - 2025-05-16
 
 ### Added